include/types.h: move shard_id_t to separate header
