+++ src/crimson/CMakeLists.txt
@@ -107,6 +107,7 @@ add_library(crimson-common STATIC
   ${PROJECT_SOURCE_DIR}/src/include/object.cc
   ${PROJECT_SOURCE_DIR}/src/include/types.cc
   ${PROJECT_SOURCE_DIR}/src/include/errorcode32.cc
+  ${PROJECT_SOURCE_DIR}/src/include/shard_id.cc
   ${PROJECT_SOURCE_DIR}/src/include/utime.cc
   ${PROJECT_SOURCE_DIR}/src/include/uuid.cc
   ${PROJECT_SOURCE_DIR}/src/librbd/Features.cc
