+++ src/test/librados/testcase_cxx.h
@@ -8,7 +8,7 @@
 #include "gtest/gtest.h"
 #include "include/rados/librados.hpp"
 #include "common/json/OSDStructures.h"
-#include "include/types.h"
+#include "include/shard_id.h"
 
 using namespace std::literals::string_view_literals;
